+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ### Other changes
 * Up `requests` lib version to 2.28.1
 * Add migging parameters to allocations.get_allocations and jobs.get_jobs (#144). Thanks @Kamilcuk
-
+* Add option for custom user agent
 ## 1.5.0
 * Add `namespace` agrument support for `get_allocations` and `get_deployments` endpoints (#133)
 * Add Python 3.10 support (#133)